pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.4"
Django = "^5.0.7"
django-imagefield = "^0.18.0"
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
format.shell = "find -type f -name '*.py' ! -path '*/migrations/*' -exec autopep8 --in-place --aggressive --aggressive '{}' \\;"
